Modularized the test functions
