+++ src/utils.rs
@@ -1,10 +1,18 @@
-use crate::constants::SEED;
 use eyre::bail;
 use ndarray::{Array1, Array2};
 use ndarray_rand::{rand_distr::Uniform, RandomExt};
 use rand::SeedableRng;
 use rand_pcg::Pcg64;
 use std::{collections::HashMap, iter::zip};
+
+/// data file
+/// NOTE: This data file has around 570k text reviews (of types: single line, paragraph).
+/// So, parse accordingly depending on the computational resources for bucketing.
+pub(crate) const DATA_FILE: &str = "./data/fine_food_reviews_1k.csv";
+
+/// seed for hyperplane generation
+/// subspace address format prefix
+pub(crate) const SEED: u64 = 2254;
 
 /// Get the embedding vector of a given text with default OpenAI embedding small model.
 /// Small embedding model: 1536 len of float values.
@@ -37,12 +45,6 @@ pub(crate) fn hash_vector(v: Vec<f64>, nbits: u16) -> String {
 	hash_str
 }
 
-#[test]
-fn test_hash_vector() {
-	let v = vec![2.3, 4.5, 6f64, 7.2];
-	assert_eq!(hash_vector(v, 4), "0010".to_string());
-}
-
 /// Distribute hashes into corresponding buckets
 pub(crate) fn bucket_hashes(v: Vec<&str>) -> HashMap<String, Vec<u128>> {
 	let mut buckets = HashMap::new();
@@ -59,26 +61,6 @@ pub(crate) fn bucket_hashes(v: Vec<&str>) -> HashMap<String, Vec<u128>> {
 	buckets
 }
 
-#[test]
-fn test_bucket_hashes() {
-	let hashed_vector = vec![
-		"1001", "1100", "1101", "1001", "1110", "1110", "1100", "1100", "1101", "1000", "1111",
-		"1111", "1101", "1100", "0100", "0100", "1111", "0100", "0010", "0100",
-	];
-	let mut expected = HashMap::new();
-	expected.insert("1001".to_string(), vec![0, 3]);
-	expected.insert("1100".to_string(), vec![1, 6, 7, 13]);
-	expected.insert("1101".to_string(), vec![2, 8, 12]);
-	expected.insert("1110".to_string(), vec![4, 5]);
-	expected.insert("1000".to_string(), vec![9]);
-	expected.insert("1111".to_string(), vec![10, 11, 16]);
-	expected.insert("0100".to_string(), vec![14, 15, 17, 19]);
-	expected.insert("0010".to_string(), vec![18]);
-
-	let result = bucket_hashes(hashed_vector.iter().map(|&s| s).collect());
-	assert_eq!(result, expected);
-}
-
 /// Calculate the Hamming distance between two strings.
 pub(crate) fn hamming_distance(str1: String, str2: String) -> eyre::Result<u16> {
 	if str1.len() != str2.len() {
@@ -93,14 +75,4 @@ pub(crate) fn hamming_distance(str1: String, str2: String) -> eyre::Result<u16> 
 	}
 
 	Ok(distance)
-}
-
-#[test]
-fn test_hamming_distance_panic() {
-	assert!(hamming_distance("0010".to_string(), "100".to_string()).is_err());
-}
-
-#[test]
-fn test_hamming_distance() {
-	assert_eq!(hamming_distance("0010".to_string(), "1001".to_string()).unwrap(), 3u16);
 }
